fix(ci): restore cross-platform integration gates

+++ crates/phantom-quic-btls/src/backend/callback_state.rs
@@ -34,7 +34,7 @@ pub(super) enum CallbackError {
         len: usize,
     },
     UnsupportedEncryptionLevel {
-        raw: u32,
+        raw: i64,
     },
     EarlyDataUnsupported,
     SecretAtInitialLevel,
